tach cac file theo router, tao file database chua db.js de chua noi dung db, chuyen doi file khoi dong index.js trong main, su dung router trong file user de tach file
