+++ views/main/index.pug
@@ -1,0 +1,2 @@
+h1 Hello
+a(href='/user/') view users
